Cleaned up API, got rid of all structs, using json instead
Made basic bot skeleton with dispatching on event types as well as
keywords for m.room.message-events. Simply needs more event types and
more keywords.

Bot uses a whitelist in channel that specifies if a user is allowed to
use the bot or not. This is stored in the channel state as
'com.severnatazvezda.apibot.user-id' and the key is:
{ "allowed": "1"/"0"}